rename firebase-config to demo-config
